Reduce allocation in Disruptor ExceptionHandler implementations
Based on PR feedback
